requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.179
botocore==1.12.179
certifi==2019.6.16
chardet==3.0.4
cymem==2.0.2
docutils==0.14
-e git+https://github.com/mobolic/facebook-sdk.git@65ff582e77f7ed68b6e9643a7490e5dee2a1031b#egg=facebook_sdk
gast==0.2.2
google-pasta==0.1.7
gpt-2-simple==0.5.3
grpcio==1.21.1
h5py==2.9.0
idna==2.8
jmespath==0.9.4
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
murmurhash==1.0.2
numpy==1.16.4
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
protobuf==3.8.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pyunsplash==1.0.0b8
regex==2019.6.8
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
spacy==2.1.4
srsly==0.0.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.0.4
toposort==1.5
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.4
wrapt==1.11.2
